refactor: CustomerServiceSelect
